% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bitfinex_OCHL_data.R
\name{bitfinex_OCHL_data}
\alias{bitfinex_OCHL_data}
\title{A function that return OCHL data from bitfinex API}
\usage{
bitfinex_OCHL_data(
  symbol = "tBTCUSD",
  TimeFrame = "1D",
  window = 10000,
  start_date = "2013-07-02 00:00:00",
  end_date = "2020-05-07 00:00:00",
  sorted = 1,
  all_data = T
)
}
\arguments{
\item{symbol}{The currency symbol that needs to have a 't' before, example: tBTCUSD, tETHUSD, etc.}

\item{TimeFrame}{The period of the candles that can be: 1D, 1h, 30m, 15m and 5m.}

\item{window}{The number of candles to be downloaded per time, maximum 10000.}

\item{start_date}{A string with start date in UTC formart 1970-01-01 00:00:00}

\item{end_date}{A string with end date in UTC format 1970-01-01 00:00:00, only use if all_data == FALSE}

\item{sorted}{A integer that can be 1 or -1, 1 means old > new}

\item{all_data}{A logical TRUE or FALSE, TRUE download all historical OCHL data, FALSE download only the window
selected with start and end date parameters}
}
\value{
A DataFrame with open, close, high, low, volume and dates
}
\description{
A function that return OCHL data from bitfinex API
}
